fixed issue with get cruelty spots service and including icon name in cruelty spot dto

+++ AFA.ServiceModel/CrueltySpot.cs
@@ -23,8 +23,10 @@ namespace AFA.ServiceModel
         public string Email { get; set; }
         public string WebpageUrl { get; set; }
 
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        //public decimal Latitude { get; set; }
+        //public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
 
         [References(typeof(CrueltySpotCategory))]
         public int CrueltySpotCategoryId { get; set; }
